feat: move Sentry init to errors package
