--- src/initializeLibraries/sentry.ts
@@ -1,14 +1,0 @@
-import * as Sentry from '@sentry/react';
-import config from 'config';
-
-import { version as APP_VERSION } from 'constants/version';
-
-const initializeSentry = () =>
-  Sentry.init({
-    dsn: config.sentryDsn,
-    environment: config.environment,
-    release: APP_VERSION,
-    attachStacktrace: true,
-  });
-
-export default initializeSentry;
